Dev to main

+++ src/main/webui/src/components/pages/+admin.svelte
@@ -1,11 +1,15 @@
+<svelte:head>
+    <title>angarium - Benutzerverwaltung</title>
+</svelte:head>
 <script>
     import {isDev, user} from "$lib/user.js";
+    import {pushPopup, popupColor} from "$lib/popup.js";
     import {onMount} from "svelte";
     import FilterList from '../shared/+filterList.svelte';
-    import Modal from '../shared/info/+modal.svelte'
+    import Modal from '../shared/info/+modal.svelte';
 
     // Icon imports ---------------------------------------
-    import {Icon} from 'svelte-icons-pack';
+    import { Icon } from 'svelte-icons-pack';
     import {TrOutlineTrashX} from "svelte-icons-pack/tr";
     import {SlReload} from "svelte-icons-pack/sl";
     import {AiOutlineUserAdd} from "svelte-icons-pack/ai";
@@ -15,8 +19,6 @@
 
     let newUserItem_exists = false;
     let newUser_name;
-
-    let showConfirm = false;
 
     let scrollElement;
     let userList = [];
@@ -85,41 +87,88 @@
                     userList = [...userList, newUser];
                     newUserItem_exists = false;
                     newUser_name = undefined;
+                    pushPopup("Der neue Benutzer wurde erfolgreich angelegt.", popupColor.success, 2500);
                 } else {
                     throw new Error("Unexpected Response: " + response.status + " (" + response.statusText + ")");
                 }
             })
             .catch((error) => {
                 console.error(error);
+                pushPopup("Beim speichern des Benutzers ist ein Fehler aufgetreten!", popupColor.error);
             });
     }
 
-    function resetPassword(userID) {
-        fetch("api/user/reset/" + userID, {
+    const ConfirmType = {
+        None: 0,
+        Reset: 1,
+        Delete: 2
+    }
+    let showResetConfirm = false;
+    let showDeleteConfirm = false;
+    let confirmType = ConfirmType.None;
+    let confirmDialog;
+    let confirmUserID = undefined;
+
+    function openConfirm(userID, type) {
+        if (type === ConfirmType.Reset) {
+            showResetConfirm = true;
+        }
+        else if (type === ConfirmType.Delete) {
+            showDeleteConfirm = true;
+        }
+        else {
+            return;
+        }
+        confirmType = type;
+        confirmUserID = userID;
+    }
+    function closeConfirm() {
+        confirmDialog.close();
+        showResetConfirm = false;
+        showDeleteConfirm = false;
+        confirmType = ConfirmType.None;
+        confirmUserID = undefined;
+    }
+
+    async function resetPassword() {
+        if (confirmUserID === undefined || confirmUserID === null || confirmUserID === "") {
+            return;
+        }
+        let id = confirmUserID;
+        closeConfirm();
+        await fetch("api/user/reset/" + id, {
             method: "POST"
         })
             .then((response) => {
-                console.log("api/user/reset/" + userID + " (POST)", response.status);
-                if (response.status === 204) {
+                console.log("api/user/reset/" + id + " (POST)", response.status);
+                if (response.status === 200) {
+                    pushPopup("Das Passwort des Benutzers wurde zurückgesetzt.", popupColor.success, 2500);
                     return null;
-                }
-                else {
+                } else {
                     throw new Error("Unexpected Response: " + response.status + " (" + response.statusText + ")");
                 }
             })
             .catch((error) => {
                 console.error(error);
+                pushPopup("Beim Zurücksetzen des Passwortes ist ein Fehler aufgetreten!", popupColor.error);
             });
+        closeConfirm();
         return null;
     }
 
-    function deleteUser(userID) {
-        fetch("api/user/" + userID, {
+    function deleteUser() {
+        if (confirmUserID === undefined || confirmUserID === null || confirmUserID === "") {
+            return;
+        }
+        let id = confirmUserID;
+        closeConfirm();
+        fetch("api/user/" + id, {
             method: "DELETE"
         })
             .then((response) => {
-                console.log("api/user/" + userID + " (DELETE)", response.status);
+                console.log("api/user/" + id + " (DELETE)", response.status);
                 if (response.status === 204) {
+                    pushPopup("Der Benutzers wurde erfolgreich gelöscht.", popupColor.success);
                     window.location.reload();
                 }
                 else {
@@ -128,6 +177,7 @@
             })
             .catch((error) => {
                 console.error(error);
+                pushPopup("Beim Löschen des Benutzers ist ein Fehler aufgetreten!", popupColor.error);
             });
         return null;
     }
@@ -137,6 +187,7 @@
             data={userList}
             field="username"
             let:item={row}
+            searchPlaceholder="Benutzername"
     >
         <header class="header" slot="header">
             <span class="id">ID</span>
@@ -158,21 +209,39 @@
                     <div class="flex-fit badge badge-accent badge-outline">admin</div>
                 {/if}
                 <div class="flex-1"></div>
-                <button class="btn btn-outline p-2 min-h-0 h-fit tooltip tooltip-bottom" data-tip="Passwort zurücksetzen" class:btn-disabled={newUserItem_exists} on:click={resetPassword(row.id)}><Icon src={SlReload} size="32" /></button>
-                <button class="btn btn-outline btn-error p-2 min-h-0 h-fit tooltip tooltip-bottom tooltip-error" data-tip="Benutzer löschen" class:btn-disabled={newUserItem_exists || row.role === "admin"} on:click={() => {/*showConfirm = true;*/ deleteUser(row.id);}}><Icon src={TrOutlineTrashX} size="32" /></button>
+                <button class="btn btn-outline p-2 min-h-0 h-fit tooltip tooltip-bottom" data-tip="Passwort zurücksetzen"
+                        class:btn-disabled={newUserItem_exists} on:click={() => {openConfirm(row.id, ConfirmType.Reset)}}>
+                    <Icon src={SlReload} size="32" />
+                </button>
+                <button class="btn btn-outline btn-error p-2 min-h-0 h-fit tooltip tooltip-bottom tooltip-error" data-tip="Benutzer löschen"
+                        class:btn-disabled={newUserItem_exists || row.role === "admin"} on:click={() => {openConfirm(row.id, ConfirmType.Delete)}}>
+                    <Icon src={TrOutlineTrashX} size="32" />
+                </button>
             {/if}
         </div>
     </FilterList>
     <div class="spacer-row h-28"></div>
 </div>
-<Modal bind:showConfirm>
-    <span slot="header">Achtung!</span>
-    <span>Wollen Sie den Benutzer wirklich löschen?</span>
-    <div class="flex flex-row gap-4">
-        <button class="btn btn-outline p-2 min-h-0 h-fit">Ja</button>
-        <button class="btn btn-outline p-2 min-h-0 h-fit">Nein</button>
-    </div>
-</Modal>
+
+{#if confirmType === ConfirmType.Reset}
+    <Modal bind:showModal={showResetConfirm} bind:dialog={confirmDialog} showClose={false}>
+        <span slot="header" class="text-lg">Achtung!</span>
+        <span>Wollen Sie das Passwort des Benutzers wirklich zurücksetzen?</span>
+        <div slot="buttons" class="flex flex-row gap-4">
+            <button class="btn btn-outline btn-accent p-4 min-h-0 h-fit min-w-20" on:click={resetPassword}>Ja</button>
+            <button class="btn btn-outline btn-error p-4 min-h-0 h-fit min-w-20" on:click={closeConfirm}>Nein</button>
+        </div>
+    </Modal>
+{:else if confirmType === ConfirmType.Delete}
+    <Modal bind:showModal={showDeleteConfirm} bind:dialog={confirmDialog} showClose={false}>
+        <span slot="header" class="text-lg">Achtung!</span>
+        <span>Wollen Sie den Benutzer wirklich löschen?</span>
+        <div slot="buttons" class="flex flex-row gap-4">
+            <button class="btn btn-outline btn-accent p-4 min-h-0 h-fit min-w-20" on:click={deleteUser}>Ja</button>
+            <button class="btn btn-outline btn-error p-4 min-h-0 h-fit min-w-20" on:click={closeConfirm}>Nein</button>
+        </div>
+    </Modal>
+{/if}
 <style>
     .header {
         @apply flex;
@@ -197,13 +266,19 @@
         @apply border-white;
     }
 
+    @media only screen and (max-width: 700px) {
+        .row {
+            @apply gap-2;
+        }
+    }
+
     .row:hover:not(header) {
         @apply border-accent;
     }
 
 </style>
 {#if $isDev}
-    <div id="dev-tools" class="fixed right-4 top-20 flex flex-col gap-2">
+    <div id="dev-tools" class="fixed right-4 top-20 flex flex-col gap-2 z-20">
         <div class="flex flex-row flex-wrap items-center justify-end gap-2">
             <button class="btn w-20" on:click={fetchUserList}>fetchRows</button>
         </div>
